square is working

+++ app/controllers/application_controller.rb
@@ -5,7 +5,7 @@ class ApplicationController < ActionController::Base
   end  
 
   def calculate_square
-    @num = params.fetch("elephant").to_i
+    @num = params.fetch("elephant").to_f
 
     @square_of_num = @num ** 2
